requirements.txt
Flask==2.3.3
gunicorn==21.2.0
loguru==0.7.2
python-dotenv==1.0.0
sentry-sdk[flask]
sentry-sdk[loguru]